Fixes to transcoding

+++ musa/transcoder.py
@@ -17,11 +17,12 @@ class TranscoderThread(MusaThread):
     """
     Class to transcode one file from Transcoder queue.
     """
-    def __init__(self,src,dst,overwrite=False):
+    def __init__(self,src,dst,overwrite=False,dry_run=False):
         MusaThread.__init__(self,'musa-transcoder')
         self.src = src
         self.dst = dst
         self.overwrite = overwrite
+        self.dry_run = dry_run
 
     def execute(self,command):
         p = subprocess.Popen(command,stdin=sys.stdin,stdout=sys.stdout,stderr=sys.stderr)
@@ -36,7 +37,10 @@ class TranscoderThread(MusaThread):
 
         if self.overwrite and os.path.isfile(self.dst.path):
             try:
-                os.unlink(self.dst.path)
+                if not self.dry_run:
+                    os.unlink(self.dst.path)
+                else:
+                    print 'Remove: %s' % self.dst.path
             except OSError,(ecode,emsg):
                 raise TranscoderError(
                     'Error removing %s: %s' % (self.dst.path,emsg)
@@ -45,7 +49,10 @@ class TranscoderThread(MusaThread):
         dst_dir = os.path.dirname(self.dst.path)
         if not os.path.isdir(dst_dir):
             try:
-                os.makedirs(dst_dir)
+                if not self.dry_run:
+                    os.makedirs(dst_dir)
+                else:
+                    print 'Create Directory: %s' % dst_dir
             except OSError,(ecode,emsg):
                 raise TranscoderError(
                     'Error creating directory %s: %s' % (dst_dir,emsg)
@@ -61,10 +68,14 @@ class TranscoderThread(MusaThread):
             raise TranscoderError(emsg)
 
         try:
-            self.status = 'decoding'
-            self.execute(decoder)
-            self.status = 'encoding'
-            self.execute(encoder)
+            if not self.dry_run:
+                self.status = 'decoding: %s' % ' '.join(decoder)
+                self.execute(decoder)
+                self.status = 'encoding: %s' % ' '.join(encoder)
+                self.execute(encoder)
+            else:
+                print 'Execute: %s' % ' '.join(decoder)
+                print 'Execute: %s' % ' '.join(encoder)
             del(wav)
         except TranscoderError,emsg:
             if wav and os.path.isfile(wav):
@@ -77,24 +88,30 @@ class TranscoderThread(MusaThread):
 
         try:
             self.status = 'tagging'
-            if self.src.tags is not None and self.dst.tags is not None:
-                self.dst.tags.update_tags(self.src.tags.as_dict())
-                if self.dst.tags.modified:
-                    self.dst.tags.save()
+            if self.src.tags is not None:
+                if not self.dry_run:
+                    if self.dst.tags is not None:
+                        self.dst.tags.update_tags(self.src.tags.as_dict())
+                        if self.dst.tags.modified:
+                            self.dst.tags.save()
+                    else:
+                        # Destination does not support tags
+                        pass
+                else:
+                    print 'Copy tags to %s' % self.dst.path
         except TagError,emsg:
             raise TranscoderError(emsg)
         self.status = 'finished'
 
 class MusaTranscoder(list):
-    def __init__(self,threads,overwrite=False):
+    def __init__(self,threads,overwrite=False,dry_run=False):
         self.threads = threads
-        self.overwrite = False
+        self.overwrite = overwrite
+        self.dry_run = dry_run
 
     def enqueue(self,src,dst):
         if not isinstance(src,Track) or not isinstance(dst,Track):
-            raise TranscoderError(
-                'Trancode source and target must be track objects'
-            )
+            raise TranscoderError('Trancode arguments must be track object')
         if not os.path.isfile(src.path):
             raise TranscoderError('No such file: %s' % src.path)
 
@@ -123,7 +140,7 @@ class MusaTranscoder(list):
                 time.sleep(0.5)
                 continue
             (src,dst) = self.pop(0)
-            t = TranscoderThread(src,dst,self.overwrite)
+            t = TranscoderThread(src,dst,self.overwrite,self.dry_run)
             t.start()
         active = threading.active_count()
         while active > 1:
